requirements.txt
python-telegram-bot==20.5
pytube==15.0.0
python-dotenv==0.21.5